abstract API call to fx
