+++ src/api/endpoints.js
@@ -1,0 +1,5 @@
+const API_BASE_URL = 'http://localhost:4001'
+
+const STUDENTS = '/students'
+
+export { API_BASE_URL, STUDENTS }
