+++ src/components/StudentCard.jsx
@@ -1,12 +1,12 @@
 import Card from 'react-bootstrap/Card'
 import React from 'react'
-import { Link } from 'react-router-dom'
 
-const StudentCard = ({ name, id }) => {
+
+const StudentCard = ({ name }) => {
   return ( <>
     <Card>
       <Card.Body>
-        <Card.Title><Link to="/students/:id">{name}</Link></Card.Title>
+        <Card.Title>{name}</Card.Title>
       </Card.Body>
     </Card>
     </>
